+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/PurchaseCreditMemoRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.PurchaseCreditMemo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +12,5 @@ import org.springframework.stereotype.Repository;
 public interface PurchaseCreditMemoRepository extends JpaRepository<PurchaseCreditMemo, String> {
 
     @Query("SELECT COALESCE(SUM(purchaseCreditMemoLine.amount), 0) FROM PurchaseCreditMemoLine purchaseCreditMemoLine WHERE purchaseCreditMemoLine.purchaseCreditMemo = :purchaseCreditMemo")
-    public double amount(@Param("purchaseCreditMemo") PurchaseCreditMemo purchaseCreditMemo);
+    public BigDecimal amount(@Param("purchaseCreditMemo") PurchaseCreditMemo purchaseCreditMemo);
 }
