+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/PurchaseOrderRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.PurchaseOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +12,5 @@ import org.springframework.stereotype.Repository;
 public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, String> {
 
     @Query("SELECT COALESCE(SUM(purchaseOrderLine.amount), 0) FROM PurchaseOrderLine purchaseOrderLine WHERE purchaseOrderLine.purchaseOrder = :purchaseOrder")
-    public double amount(@Param("purchaseOrder") PurchaseOrder purchaseOrder);
+    public BigDecimal amount(@Param("purchaseOrder") PurchaseOrder purchaseOrder);
 }
