+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/SalesCreditMemoRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.SalesCreditMemo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,6 +12,6 @@ import org.springframework.stereotype.Repository;
 public interface SalesCreditMemoRepository extends JpaRepository<SalesCreditMemo, String> {
 
     @Query("SELECT COALESCE(SUM(salesCreditMemoLine.amount), 0) FROM SalesCreditMemoLine salesCreditMemoLine WHERE salesCreditMemoLine.salesCreditMemoLineId.salesCreditMemo = :salesCreditMemo")
-    public double amount(@Param("salesCreditMemo") SalesCreditMemo salesCreditMemo);
+    public BigDecimal amount(@Param("salesCreditMemo") SalesCreditMemo salesCreditMemo);
 
 }
