+++ erp.backend/src/main/java/org/roko/erp/backend/repositories/SalesOrderRepository.java
@@ -1,5 +1,7 @@
 package org.roko.erp.backend.repositories;
 
+import java.math.BigDecimal;
+
 import org.roko.erp.backend.model.SalesOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +12,5 @@ import org.springframework.stereotype.Repository;
 public interface SalesOrderRepository extends JpaRepository<SalesOrder, String> {
 
     @Query("SELECT COALESCE(SUM(salesOrderLine.amount), 0) FROM SalesOrderLine salesOrderLine WHERE salesOrderLine.salesOrderLineId.salesOrder = :salesOrder")
-    public double amount(@Param("salesOrder")SalesOrder salesOrder);
+    public BigDecimal amount(@Param("salesOrder")SalesOrder salesOrder);
 }
